Time: 3 ms (39.35%) | Memory: 16.9 MB (84.97%) - LeetSync

+++ 150-evaluate-reverse-polish-notation/evaluate-reverse-polish-notation.cpp
@@ -36,7 +36,7 @@ public:
         }
         i++;
       }
-      long long ans=q.top();
+      int ans=q.top();
       return ans;  
     }
 };
